zsh: update CDPATH
Sometimes it's hard to figure out where exactly in the GOPATH a
repository should live when there is not necessarily one canonical
remote. Now that Go supports modules outside of GOPATH, I think I'd like
to move away from keeping everything in code/src/ and simply keep
repositories in a top-level directory. Eventually, I think I'll want to
rename the directories:

    code -> go
    c -> code

But as this is relatively experimental for me currently, I'm holding off
on that and just updating CDPATH.